requirements.txt
langchain
langchain_community
fastapi
uvicorn
langserve
sse_starlette 
streamlit
python-dotenv
tavily-python
langchain_groq